Implement function signature checking

+++ tests/test_procedures.py
@@ -6,7 +6,7 @@ def test_simple_procedure_definition():
 import wasi_snapshot_preview1::proc_exit(exit_code: u32) as proc_exit;
 
 fn _start() {
-    proc_exit(2)
+    proc_exit(2u32)
 }"""
     exit_code, _ = tools.run_test_code(code)
 
@@ -17,8 +17,8 @@ def test_function_without_parameters_with_return_type_is_usable():
     code = """
 import wasi_snapshot_preview1::proc_exit(exit_code: u32) as proc_exit;
 
-fn bla() -> s32 {
-    2 + 3
+fn bla() -> u32 {
+    (2 + 3) as u32
 }
 
 fn _start() {
@@ -34,11 +34,11 @@ def test_multiple_function_definitions():
 import wasi_snapshot_preview1::proc_exit(exit_code: u32) as proc_exit;
 
 fn foo() {
-    proc_exit(2)
+    proc_exit(2u32)
 }
 
 fn _start() {
-    proc_exit(2)
+    proc_exit(2u32)
 }"""
     exit_code, _ = tools.run_test_code(code)
 
@@ -54,7 +54,7 @@ fn foo(a: s32, b: s32) -> s32 {
 }
 
 fn _start() {
-    proc_exit(foo(1, 2))
+    proc_exit(foo(1, 2) as u32)
 }"""
     exit_code, _ = tools.run_test_code(code)
 
@@ -70,7 +70,7 @@ fn add2(a: u32) -> u32 {
 }
 
 fn _start() {
-    proc_exit(add2(3))
+    proc_exit(add2(3u32))
 }"""
     exit_code, _ = tools.run_test_code(code)
